Add inline documentation for public functions
